pyproject.toml
[tool.poetry]
name = "game-lib-project"
version = "0.1.0"
description = "Portfolio project. GameLib is a video game store site!"
authors = ["Aurun <aurunchill@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.11"
python-dotenv = "^1.0.1"
social-auth-app-django = "^5.4.1"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
